duration displaying on list items, song icon on player control view, main menu algorithm update

+++ app/src/main/java/com/nafanya/mp3world/model/listManagers/MediaStoreReader.kt
@@ -72,7 +72,7 @@ object MediaStoreReader {
                 val thisAlbumId = cursor.getLong(albumIdColumn)
                 val thisAlbumName = cursor.getString(albumColumn)
                 val thisPath = cursor.getString(pathColumn)
-                val thisDuration = cursor.getInt(durationColumn)
+                val thisDuration = cursor.getLong(durationColumn)
                 // tracks with <unknown> artist are corrupted
                 if (thisArtist != "<unknown>") {
                     // set the song art
